+++ src/afterglow/controllers/tempo.clj
@@ -100,7 +100,10 @@
   next higher level. (In other words, if it was going to be a tempo
   tap, it would be treated as a beat tap; what would normally be a
   beat tap would be treated as a bar tap, and a bar tap would be
-  promoted to start a phrase.)
+  promoted to start a phrase.) When you map a shift button
+  using [[add-midi-control-to-shift-mapping]], it gives you a function
+  that is suitable for use with `:shift-fn` as the value of the
+  `:state` key in the map it returns.
 
   Returns the tempo-mapping function which can be passed
   to [[remove-midi-control-to-tempo-mapping]] if you ever want to stop
@@ -111,7 +114,8 @@
   {:pre [(have? some? *show*) (have? #{:control :note} kind) (have? some? device-filter)
          (have? integer? channel) (have? #(<= 0 % 15) channel) (have? integer? note) (have? #(<= 0 % 127) note)
          (have? #(or (not %) (and (integer? %) (<= 0 % 127))) feedback-on)
-         (have? integer? feedback-off) (have? #(<= 0 % 127) feedback-off)]}
+         (have? integer? feedback-off) (have? #(<= 0 % 127) feedback-off)
+         (have? fn? shift-fn)]}
   (let [feedback-device (when feedback-on (amidi/find-midi-out device-filter))
         tap-handler (create-show-tempo-tap-handler *show* :shift-fn shift-fn)
         midi-handler (fn [message]
@@ -123,7 +127,7 @@
     (case kind
       :control (amidi/add-control-mapping device-filter channel note midi-handler)
       :note (amidi/add-note-mapping device-filter channel note midi-handler))
-    tap-handler))
+    midi-handler))
 
 (defn remove-midi-control-to-tempo-mapping
   "Undoes the effect of [[add-midi-control-to-tempo-mapping]]. You
@@ -140,3 +144,73 @@
       :control (amidi/remove-control-mapping device-filter channel note f)
       :note (amidi/remove-note-mapping device-filter channel note f))
     nil))
+
+(defn add-midi-control-to-shift-mapping
+  "Sets up a controller pad or button to act as a shift button which
+  can be used to modify other mappings. The shift state is entered
+  whenever the specified note (when `kind` is `:note`) or
+  controller-change (when `kind` is `:control`) message is received
+  with a non-zero velocity (or control value), and left whenever a
+  matching note-off or controller message with velocity zero is
+  received.
+
+  The device to be mapped is identified by `device-filter`. The first
+  input port which matches using [[filter-devices]] will be used.
+
+  Afterglow will attempt to provide feedback by lighting the pad or
+  button when it is held down by sending note on/off or control-change
+  values to the same controller. The note velocities or control values
+  used can be changed by passing in different values with
+  `:feedback-on` and `:feedback-off`, and this behavior can be
+  suppressed entirely by passing `false` with `:feedback-on`.
+
+  Returns a map containing an entry `:state` whose value is a function
+  that can be called to check the state of the shift button, returning
+  `true` whenever it is held down. There are other entries in the map
+  which are used for the mapping's own purposes. Pass the entire map
+  to [[remove-midi-control-to-shift-mapping]] if you ever want to stop
+  the MIDI control or note from acting like a shift button."
+  [device-filter channel kind note & {:keys [feedback-on feedback-off]
+                                      :or {feedback-on 127 feedback-off 0}}]
+  {:pre [(have? #{:control :note} kind) (have? some? device-filter)
+         (have? integer? channel) (have? #(<= 0 % 15) channel) (have? integer? note) (have? #(<= 0 % 127) note)
+         (have? #(or (not %) (and (integer? %) (<= 0 % 127))) feedback-on)
+         (have? integer? feedback-off) (have? #(<= 0 % 127) feedback-off)]}
+  (let [feedback-device (when feedback-on (amidi/find-midi-out device-filter))
+        state (atom false)
+        midi-handler (fn [message]
+                       (reset! state (and (pos? (:velocity message))
+                                          (or (= :control kind)
+                                              (= :note-on (:command message)))))
+                       (when feedback-device
+                         (if (= :control kind)
+                           (midi/midi-control feedback-device note (if @state feedback-on feedback-off) channel)
+                           (if @state
+                             (midi/midi-note-on feedback-device note feedback-on channel)
+                             (midi/midi-note-off feedback-device note channel)))))]
+    (case kind
+      :control (amidi/add-control-mapping device-filter channel note midi-handler)
+      :note (amidi/add-note-mapping device-filter channel note midi-handler))
+    (with-meta
+      {:state (fn [] @state)
+       :handler midi-handler
+       :clear-handler (fn []
+                        (when feedback-device
+                          (if (= :control kind)
+                            (midi/midi-control feedback-device note feedback-off channel)
+                            (midi/midi-note-off feedback-device note channel))))}
+      {:type :midi-shift-mapping})))
+
+(defn remove-midi-control-to-shift-mapping
+  "Undoes the effect of [[add-midi-control-to-tempo-mapping]]. You
+  must pass the value that was returned by that function as the
+  argument `mapping`."
+  [device-filter channel kind note mapping]
+  {:pre [(have? #{:control :note} kind) (have? some? device-filter)
+         (have? integer? channel) (have? #(<= 0 % 15) channel) (have? integer? note) (have? #(<= 0 % 127) note)
+         (have? #(= (type %) :midi-shift-mapping) mapping)]}
+  (case kind
+    :control (amidi/remove-control-mapping device-filter channel note (:handler mapping))
+    :note (amidi/remove-note-mapping device-filter channel note (:handler mapping)))
+  ((:clear-handler mapping))
+  nil)